+++ stationlink/stream_test.go
@@ -11,9 +11,9 @@ import (
 	"github.com/macula-io/macula-go/cbor"
 	"github.com/macula-io/macula-go/frame"
 	"github.com/macula-io/macula-go/identity"
-	"github.com/macula-io/macula-go/internal/teststation"
 	"github.com/macula-io/macula-go/profile"
 	"github.com/macula-io/macula-go/stationlink"
+	"github.com/macula-io/macula-go/teststation"
 )
 
 const streamProcedure = "mcl-tube/watch"